Move `PromotedType<T>` to `:utility`
It's a pretty generic utility, and it'll be handy to have it in a more
common location.

Helps pave the way for #349.
