+++ au/apply_rational_magnitude_to_integral_test.cc
@@ -36,24 +36,6 @@ constexpr void ensure_relevant_kind_of_magnitude(Magnitude<BPs...> m) {
     static_assert(!is_integer(ONE / m), "Magnitude must not be purely inverse-integer");
 }
 
-TEST(PromotedType, IdentityForInt) {
-    // `int` does not undergo type promotion.
-    StaticAssertTypeEq<int, PromotedType<int>>();
-}
-
-TEST(PromotedType, PromotesUint8TIntoLargerType) {
-    using PromotedU8 = PromotedType<uint8_t>;
-
-    // Technically, this need not be true on every conceivable architecture.  However, it is true on
-    // the vast majority that are used in practice.  Moreover, the failure mode if it's not is
-    // simply that a test would fail when run on some obscure architecture, and the failure would
-    // direct the user to this comment.  This doesn't affect the actual library usage one way or
-    // another.
-    ASSERT_THAT((std::is_same<uint8_t, PromotedU8>::value), IsFalse());
-
-    EXPECT_THAT(sizeof(PromotedU8), Gt(sizeof(uint8_t)));
-}
-
 TEST(IsAbsKnownToBeLessThanOne, ProducesExpectedResultsForMagnitudesThatCanFitInUintmax) {
     EXPECT_THAT(is_abs_known_to_be_less_than_one(mag<1>() / mag<2>()),
                 Eq(IsAbsMagLessThanOne::DEFINITELY));
